+++ NEWS.md
@@ -4,6 +4,10 @@ GeothermalWells.jl follows [semantic versioning](https://semver.org/).
 
 ## Unreleased
 
+### Added
+- Added Brown et al. validation data loaders for Figure 6 array cases: `data_brown_array(spacing, i)`  
+- `data_brown_single_well_c` had the wrong name and is now `data_brown_single_well_d`, but `data_brown_single_well_c` is still there for legacy-compatible.
+
 ## v0.3.0
 
 ### Added